8/25 17:57 添加了选课操作以及token等

+++ src/main/java/com/test/trainingprogrambackend/mapper/CourseMapper.java
@@ -11,14 +11,16 @@ import java.util.List;
 
 @Mapper
 public interface CourseMapper {
+    //选择所有课程
     @Select("select * from course")
     List<Course> selectAll();
-
+    //通过课程名字选择课程
     @Select("select * from course where name=#{name}")
     List<Course> getCourseByName(String name);
+    //通过课程编码选择课程
     @Select("select * from course where code=#{code}")
     List<Course> getCourseByCode(String code);
-
+    //通过课程名称关键字选择课程
     @Select("select * from course where name like concat('%',#{name},'%')")
     List<Course> getCourseByNameLike(@Param("name") String name);
 
@@ -34,9 +36,10 @@ public interface CourseMapper {
             +"<if test='selectedNumber!=null'> AND selectedNumber=#{selectedNumber}</if>"
             +"<if test='isFull!=null'> AND isFull=#{isFull}</if>"
             +"</script>")
-    List<CourseClassDTO> selectByConditions(String code, String name, Integer credit, String type, String deptName, String classNumber, String teacherName, Integer capacity, Integer selectedNumber, Integer isFull);
+    List<CourseClassDTO> selectByConditions(String code, String name, Float credit, String type, String deptName, String classNumber, String teacherName, Integer capacity, Integer selectedNumber, Integer isFull);
 
     @Select("select * from course natural join courseclass natural join takes where studentid=#{studentid}")
     List<CourseClassDTO> courseClassByStudentId(@Param("studentid") String studentid);
+
 
 }
